Apply standard Python project layout
This is an attempt to make the *py-optics* project conform to Python
packaging standards [0-2]:

- Add a `pyproject.toml` file. It doesn't do much at this point, besides
declaring its dependency on *setuptools* for building.

- Use a more standard directory layout, based on a `src/` root directory
containing the various packages [1]. There are definitely other ways to
approach this, but the arguments in favor of a `src/` directory seem to
make sense. Example notebooks go into a top-level `examples/` directory.

- Use a `setup.cfg` file for declaring package metadata. We apparently
still need a stub `setup.py` file [0] to make it possible to use `pip
install -e .`.

- Move code into a package, with a top-level *pyoptics* package that
contains the other subpackages. The top-level *pyoptics* package
contains appropriate metadata, including a title, author, and a `0.1.0`
version number. Subpackages have their main code in the `__init__`
module (again, there are other ways to do this, but in this case it
seems sensible to do it this way).

- Apply uniform formatting of docstrings.

- Add "Getting started" section to README.

- Add LICENSE file.

 [0]: https://snarky.ca/what-the-heck-is-pyproject-toml/
 [1]: https://hynek.me/articles/testing-packaging/
[2]:
https://setuptools.readthedocs.io/en/latest/userguide/declarative_config.html